+++ src/ui/react/weg/app.tsx
@@ -1,7 +1,7 @@
 import { ErrorBoundary } from "./components/Error/index.tsx";
 import { SeelenWeg } from "./modules/bar/index.tsx";
 import { useSignalEffect } from "@preact/signals";
-import { $isSomeFullscreenOnMonitor } from "./modules/shared/state/windows.ts";
+import { $top_interactable_window } from "./modules/shared/state/windows.ts";
 import { Widget } from "@seelen-ui/lib";
 import { $initialPositionSet } from "./modules/shared/state/settings.ts";
 import { useEffect } from "preact/hooks";
@@ -12,7 +12,7 @@ export function App() {
   }, []);
 
   useSignalEffect(() => {
-    const fullscreened = $isSomeFullscreenOnMonitor.value;
+    const fullscreened = $top_interactable_window.value?.isFullscreen;
     if (fullscreened) {
       Widget.self.window.setAlwaysOnTop(false);
       Widget.self.window.setAlwaysOnBottom(true);